packages/jupyter-ai/jupyter_ai/mcp_rag/requirements.txt
faiss-cpu
sentence-transformers
flask
requests
numpy
